XMLHttpRequest/send-no-response-event-loadstart.htm

<!DOCTYPE html>
<html>
<head>
    <link rel="help" href="http://dvcs.w3.org/hg/xhr/raw-file/tip/Overview.html#the-send()-method" />
    <script src="/resources/testharness.js"></script>
    <title>XMLHttpRequest L2: The send() method: Fire a progress event named loadstart (no response entity body and the state is LOADING)</title>
</head>

<body>
    <div id="log"></div>
    
    <script type="text/javascript">
        var test = async_test();

        test.step(function()
        {
            var xhr = new XMLHttpRequest();

            xhr.onreadystatechange = function() 
            {
                test.step(function() 
                {
                    if (xhr.readyState == 3) 
                    {
                        assert_equals(xhr.response, "");
                    }
                    else if (xhr.readyState == 4) 
                    {
                        assert_unreached("loadstart event did not fire in LOADING state!");
                    }
                });
            };

            xhr.onloadstart = function() 
            {
                test.step(function() { test.done("Test done!"); });
            };

            xhr.open("POST", "./support/echo.php", true);
            xhr.send();
        });
    </script>
</body>
</html>
